+++ rtic-sync/CHANGELOG.md
@@ -9,6 +9,7 @@ For each category, _Added_, _Changed_, _Fixed_ add new entries at the top!
 
 ### Fixed
 
+- A `Channel<T>` is safely `Send` and `Sync` when `T: Send`.
 - An `Arbiter<T>` is safely `Send` and `Sync` when `T: Send`.
 
 ## v1.5.0 - 2026-05-30